Simplified callback mechanisms.
Each button would call a unique method to display or close the required window.  These have been replaced by a much smaller number of callback functions that take a parameter, and use that parameter to decide which windows to close/open.

Firstly, there is now one callback to hide windows (hide_parent), instead of a separate one for each window with a close button.

Secondly, the dispaying of the winning divisions is done by one callback which takes a parameter.

Thirdly, the displaying of the number entry form is done by one callback which takes a parmeter.

As this was the first proper GUI program I made using FLTK, the original strategy what to do what just works.  This has been refactored to a much simpler, more extensible and maintanable state.